# openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Script Labs API",
    "description": "A secure Script Labs API built with Node.js and Express, featuring JWT authentication via Supabase and CRUD operations for labs management.",
    "version": "1.0.0",
    "contact": {
      "name": "Hendri Christianto",
      "email": "hendri.christianto24@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://script-labs-app-z8p8.onrender.com",
      "description": "Production server on Render"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/labs/search": {
      "get": {
        "summary": "Search labs by title or author",
        "description": "Search labs for the authenticated user by title or author. Tidak mengubah database.",
        "tags": ["Labs"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search query (title or author)",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default: 1)",
            "required": false,
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page (default: 10, max: 100)",
            "required": false,
            "schema": { "type": "integer", "default": 10, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Lab" }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": { "type": "integer", "example": 1 },
                        "limit": { "type": "integer", "example": 10 },
                        "total": { "type": "integer", "example": 2 },
                        "totalPages": { "type": "integer", "example": 1 }
                      }
                    },
                    "search_query": { "type": "string", "example": "atomic" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": 14,
                      "title": "Atomic Habits",
                      "author": "James Clear",
                      "user_id": "cd004999d-7d4b-4167-a9af-0f15783caee5"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "limit": 10,
                    "total": 1,
                    "totalPages": 1
                  },
                  "search_query": "atomic",
                  "timestamp": "2025-07-31T10:00:00.000Z"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/labs": {
      "get": {
        "summary": "Get all labs",
        "description": "Retrieve all labs belonging to the authenticated user",
        "tags": ["Labs"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of labs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lab"
                  }
                },
                "example": [
                  {
                    "id": 14,
                    "title": "buku menari",
                    "author": "jojo",
                    "user_id": "cd004999d-7d4b-4167-a9af-0f15783caee5"
                  },
                  {
                    "id": 15,
                    "title": "buku mewarnai",
                    "author": "santi",
                    "user_id": "cd004999d-7d4b-4167-a9af-0f15783caee5"
                  }
                ]
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "summary": "Add a new lab",
        "description": "Create a new lab entry for the authenticated user",
        "tags": ["Labs"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookInput"
              },
              "example": {
                "title": "Atomic Habits",
                "author": "James Clear"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "lab created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/lab"
                },
                "example": {
                  "id": 25,
                  "title": "Atomic Habits",
                  "author": "James Clear",
                  "user_id": "cd004999d-7d4b-4167-a9af-0f15783caee5"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Title and author are required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/labs/{id}": {
      "put": {
        "summary": "Update a lab",
        "description": "Update an existing lab (user can only update their own labs)",
        "tags": ["labs"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "lab ID",
            "schema": {
              "type": "integer",
              "example": 14
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookInput"
              },
              "example": {
                "title": "Atomic Habits Updated",
                "author": "J. Clear"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "lab updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/lab"
                },
                "example": {
                  "id": 14,
                  "title": "Atomic Habits Updated",
                  "author": "J. Clear",
                  "user_id": "cd004999d-7d4b-4167-a9af-0f15783caee5"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "All fields are required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "lab not found or unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "lab not found or unauthorized"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete a lab",
        "description": "Delete an existing lab (user can only delete their own labs)",
        "tags": ["labs"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "lab ID",
            "schema": {
              "type": "integer",
              "example": 14
            }
          }
        ],
        "responses": {
          "200": {
            "description": "lab deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                },
                "example": {
                  "message": "lab deleted successfully"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "lab not found or unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "lab not found or unauthorized"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from Supabase authentication"
      }
    },
    "schemas": {
      "lab": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the lab",
            "example": 14
          },
          "title": {
            "type": "string",
            "description": "Title of the lab",
            "example": "Atomic Habits"
          },
          "author": {
            "type": "string",
            "description": "Author of the lab",
            "example": "James Clear"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user who owns this lab",
            "example": "cd004999d-7d4b-4167-a9af-0f15783caee5"
          }
        },
        "required": ["id", "title", "author", "user_id"]
      },
      "BookInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the lab",
            "example": "Atomic Habits"
          },
          "author": {
            "type": "string",
            "description": "Author of the lab",
            "example": "James Clear"
          }
        },
        "required": ["title", "author"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "An error occurred"
          },
          "error": {
            "type": "string",
            "description": "Detailed error information"
          }
        },
        "required": ["message"]
      },
      "SuccessMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Operation completed successfully"
          }
        },
        "required": ["message"]
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication information is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "no_token": {
                "summary": "No token provided",
                "value": {
                  "message": "No token provided"
                }
              },
              "invalid_token": {
                "summary": "Invalid token",
                "value": {
                  "message": "Invalid token"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "message": "Server error saat menambah buku",
              "error": "Database connection failed"
            }
          }
        }
      }
    }
  },
...
}